+++ packages/leemons/src/calendar/CalendarSubNavFilters/CalendarSubNavFilters.styles.js
@@ -1,55 +1,22 @@
 import { createStyles, getFontExpressive } from '@bubbles-ui/components';
 
-export const CalendarSubNavFiltersStyles = createStyles((theme, { mainColor, lightMode }) => {
+export const CalendarSubNavFiltersStyles = createStyles((theme, {}) => {
+  const menuTheme = theme.other.menu ?? {};
+  console.log('menuTheme:', menuTheme);
   return {
-    subNav: {
+    root: {
       height: '100%',
+      backgroundColor: menuTheme.background?.color?.main?.default,
+      borderRight: `1px solid ${menuTheme.border?.color?.main?.default}`,
+      padding: menuTheme.spacing.gap,
     },
-    switchLabel: {
-      color: lightMode ? theme.colors.text01 : theme.colors.text07,
-      fontWeight: 500,
+    subtitle: {
+      ...(menuTheme.content ?? {})["typo--medium"],
+      color: menuTheme.content?.color?.main?.default
     },
-    segmentRoot: {
-      height: '42px',
-      padding: 2,
-      backgroundColor: !lightMode && mainColor,
-      // paddingLeft: 3,
-      // paddingRight: 0,
-    },
-    segmentRoot2: {
-      width: '100%',
-      padding: 2,
-      backgroundColor: !lightMode && mainColor,
-    },
-    segmentLabel: {
-      ...getFontExpressive(),
-      color: lightMode ? theme.colors.text05 : theme.colors.text06,
-      boxSizing: 'border-box',
-      alignItems: 'center',
-      textAlign: 'center',
-      display: 'flex',
-      height: '100%',
-      margin: 0,
-      '&:hover': {
-        color: !lightMode && theme.colors.text07,
-      },
-    },
-    segmentLabelActive: {
-      color: lightMode ? `${theme.colors.text01}!important` : `${theme.colors.text07}!important`,
-    },
-    segmentActive: {
-      backgroundColor: lightMode ? theme.colors.mainWhite : 'rgba(255, 255, 255, 0.25)',
-      // backgroundColor: theme.colors.interactive02d,
-      top: 4,
-      left: 4,
-    },
-    segmentControl: {
-      border: 'none!important',
-    },
-    icon: {
-      img: {
-        filter: 'brightness(0) invert(1)',
-      },
-    },
+    switchWrapper: {
+      marginLeft: -8,
+      marginBottom: -4
+    }
   };
 });